Ported over the notebooks to Python. Still missing Planner and bug with TextMemory
